refactor backend code
