+++ backend/routes/getAlbumById.js
@@ -1,0 +1,22 @@
+import express from "express";
+import axios from "axios";
+
+const router = express.Router();
+
+router.get("/", async (req, res) => {
+  const id = req.query.id;
+  console.log("Song id is " + id);
+  const apiUrl = `https://saavn.dev/api/albums?id=${encodeURIComponent(id)}`;
+  try {
+    const respone = await axios.get(apiUrl);
+    if (respone.status != 200) {
+      throw new Error("Failed to fetch data from the external API");
+    } else {
+      res.status(200).json(respone.data);
+    }
+  } catch (error) {
+    res.status(500).json({ error: "Internal server error" });
+  }
+});
+
+export default router;
